+++ zcash_primitives/src/sapling/verifier.rs
@@ -1,7 +1,8 @@
 use bellman::{gadgets::multipack, groth16::Proof};
 use bls12_381::Bls12;
 use group::{ff::PrimeField, Curve, GroupEncoding};
-use zcash_primitives::{
+
+use crate::{
     sapling::{
         note::ExtractedNoteCommitment,
         redjubjub::{PublicKey, Signature},