+++ zcash_proofs/src/sapling/mod.rs
@@ -1,6 +1,3 @@
 //! Helpers for creating Sapling proofs.
 
 mod prover;
-mod verifier;
-
-pub use self::verifier::{BatchValidator, SaplingVerificationContext};